relocated SimPapasDetector to inside SimPapas
