+++ Sim/SimPapas/CMakeLists.txt
@@ -4,15 +4,17 @@
 gaudi_subdir(SimPapas v1r0)
 
 # this declaration will not be needed in the future
-gaudi_depends_on_subdirs(GaudiAlg Sim/SimG4Common Sim/SimPapasDetector)
+gaudi_depends_on_subdirs(GaudiAlg Sim/SimG4Common FWCore GaudiKernel)
 
 find_package(Geant4)
 include(${Geant4_USE_FILE})
 
 find_package(FCCEDM)
 find_package(PODIO)
-find_package(ROOT COMPONENTS Tree Physics EG)
+find_package(ROOT COMPONENTS MathCore GenVector Tree Physics EG)
 find_package(papas)
+find_package(Geant4)
+include(${Geant4_USE_FILE})
 
 gaudi_add_library(PapasUtils
                  src/lib/*.cpp
@@ -23,10 +25,21 @@ gaudi_add_library(PapasUtils
 
 gaudi_add_module(SimPapas
                  src/*.cpp
-                 INCLUDE_DIRS FWCore ROOT FCCEDM PODIO papas SimG4Common SimPapasDetector
-                 LINK_LIBRARIES GaudiAlgLib ROOT FCCEDM PODIO ROOT papas)
+                 INCLUDE_DIRS  GaudiKernel ROOT Geant4 FWCore FCCEDM PODIO papas SimG4Common SimPapasDetector
+                 LINK_LIBRARIES GaudiKernel Geant4  GaudiAlgLib ROOT FCCEDM PODIO ROOT papas)
 
+gaudi_add_module(CMSDetector
+                 src/CMSDetector/*.cpp
+                 INCLUDE_DIRS  GaudiKernel ROOT Geant4 FWCore FCCEDM PODIO papas SimG4Common SimPapasDetector
+                 LINK_LIBRARIES GaudiKernel Geant4  GaudiAlgLib ROOT FCCEDM PODIO ROOT papas)
+
+gaudi_add_module(ClicDetector
+                 src/ClicDetector/*.cpp
+                 INCLUDE_DIRS  GaudiKernel ROOT Geant4 FWCore FCCEDM PODIO papas SimG4Common SimPapasDetector
+                 LINK_LIBRARIES GaudiKernel Geant4  GaudiAlgLib ROOT FCCEDM PODIO ROOT papas)
 gaudi_install_headers(SimPapas)
+gaudi_install_headers(CMSDetector)
+gaudi_install_headers(ClicDetector)
 gaudi_install_python_modules()
 
 
@@ -34,3 +47,4 @@ include(CTest)
 gaudi_add_test(simple_papastool
                WORKING_DIRECTORY ${PROJECT_SOURCE_DIR}
                FRAMEWORK options/simple_papastool.py)
+
