+++ Sim/SimPapas/src/CMSDetector/CMSFieldSvc.h
@@ -1,7 +1,7 @@
 #ifndef CMSFIELDSVC_H
 #define CMSFIELDSVC_H
 
-#include "SimPapasDetector/IPapasFieldSvc.h"
+#include "SimPapas/IPapasFieldSvc.h"
 
 #include "GaudiKernel/MsgStream.h"
 #include "GaudiKernel/Service.h"
