+++ Sim/SimPapas/src/ClicDetector/ClicDetectorSvc.cpp
