+++ Sim/SimPapas/src/ClicDetector/ClicEcalSvc.cpp
