+++ Sim/SimPapas/src/ClicDetector/ClicFieldSvc.cpp
