+++ Sim/SimPapas/src/ClicDetector/ClicHcalSvc.h
@@ -1,7 +1,7 @@
 #ifndef CLICHCALSVC_H
 #define CLICHCALSVC_H
 
-#include "SimPapasDetector/IPapasCalorimeterSvc.h"
+#include "SimPapas/IPapasCalorimeterSvc.h"
 
 #include "GaudiKernel/MsgStream.h"
 #include "GaudiKernel/Service.h"
